KTOR-730 Move existing tests into their packages
