+++ ktor-client/ktor-client-core/common/test/io/ktor/client/TestUtils.kt
@@ -1,12 +1,14 @@
+/*
+ * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client
+
 import io.ktor.client.engine.*
 import io.ktor.client.request.*
 import io.ktor.util.*
 import kotlinx.coroutines.*
 import kotlin.coroutines.*
-
-/*
- * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
 
 internal object TestEngine: HttpClientEngine {
     override val dispatcher: CoroutineDispatcher
@@ -25,4 +27,3 @@ internal object TestEngine: HttpClientEngine {
     override fun close() {
     }
 }
-
