+++ ktor-client/ktor-client-core/common/test/io/ktor/client/features/cache/CacheExpiresTest.kt
@@ -1,7 +1,8 @@
 /*
  * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
-package io.ktor.client.features.cache.tests
+
+package io.ktor.client.features.cache
 
 import io.ktor.client.call.*
 import io.ktor.client.features.cache.*
