+++ ktor-client/ktor-client-core/common/test/io/ktor/client/features/cookies/CookiesTest.kt
@@ -1,12 +1,13 @@
-import io.ktor.client.features.cookies.*
+/*
+ * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client.features.cookies
+
 import io.ktor.client.request.*
 import io.ktor.http.*
 import io.ktor.test.dispatcher.*
 import kotlin.test.*
-
-/*
- * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
 
 class CookiesTest {
