+++ ktor-client/ktor-client-core/common/test/io/ktor/client/request/forms/MultiPartFormDataContentTest.kt
@@ -1,12 +1,13 @@
-import io.ktor.client.request.forms.*
+/*
+ * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client.request.forms
+
 import io.ktor.test.dispatcher.*
 import io.ktor.utils.io.*
 import io.ktor.utils.io.core.*
 import kotlin.test.*
-
-/*
- * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
 
 class MultiPartFormDataContentTest {
